+++ Server/bin/pom.xml
@@ -98,7 +98,7 @@
                 <archive>
                     <manifest>
                         <addClasspath>true</addClasspath>
-                        <mainClass>com.movies.Application</mainClass> <!-- Replace with your actual main class -->
+                        <mainClass>Application</mainClass> <!-- Replace with your actual main class -->
                     </manifest>
                 </archive>
             </configuration>
