+++ Server/bin/src/main/resources/application.properties
@@ -1,0 +1,15 @@
+# Database Configuration
+spring.datasource.url=jdbc:mysql://localhost:3306/movies_DB
+spring.datasource.username=mark
+spring.datasource.password=password
+
+# Hibernate specific configurations
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5Dialect
+spring.jpa.hibernate.ddl-auto=update
+
+# Logging Configuration
+logging.level.org.springframework=INFO
+logging.level.com.example=DEBUG
+
+# Server Configuration
+server.port=3001
