--- Server/main/java/config/connection.java
@@ -1,5 +1,0 @@
-package config;
-
-public class connection {
-    
-}
