--- Server/main/java/repositories/UserRepository.java
@@ -1,13 +1,0 @@
-package repositories;
-
-import models.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    User findByUsername(String username);
-
-    // You can add additional custom query methods here if needed
-}
