+++ Server/src/main/java/com/movies/config/webConfig.java
@@ -1,0 +1,15 @@
+package com.movies.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class webConfig implements WebMvcConfigurer {
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+            .allowedOrigins("http://localhost:3000") // Replace with your front-end's URL
+            .allowedMethods("GET", "POST", "PUT", "DELETE");
+    }
+}
