+++ Server/src/main/java/com/movies/models/AuthResponse.java
@@ -1,0 +1,10 @@
+package com.movies.models;
+
+public class AuthResponse {
+    private String token;
+    private User user;
+    public AuthResponse(String token, User user) {
+        this.token = token;
+        this.user = user;
+    }   
+}
