+++ Server/src/main/java/com/movies/repositories/UserRepository.java
@@ -1,0 +1,13 @@
+package com.movies.repositories;
+
+import com.movies.models.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    User findByUsername(String username);
+
+    // You can add additional custom query methods here if needed
+}
